Module fil_actor_verifreg_state::v9

Re-exports

pub use self::state::Allocation;

pub use self::state::Claim;

pub use self::state::State;

Modules

state

Structs

AddrPairKey

AllocationRequest

AllocationRequests

Operator-data payload for a datacap token transfer receiver hook specifying an allocation. The implied client is the sender of the datacap.

AllocationsResponse

Recipient data payload in response to a datacap token transfer.

ClaimAllocationsParams

ClaimAllocationsReturn

ClaimExtensionRequest

ClaimTerm

ExtendClaimTermsParams

GetClaimsParams

GetClaimsReturn

RemoveDataCapParams

RemoveDataCapProposal

RemoveDataCapProposalID

RemoveDataCapRequest

RemoveDataCapReturn

RemoveExpiredAllocationsParams

RemoveExpiredAllocationsReturn

RemoveExpiredClaimsParams

RemoveExpiredClaimsReturn

SectorAllocationClaim

VerifierParams

Enums

Method

Account actor methods available

Constants

SIGNATURE_DOMAIN_SEPARATION_REMOVE_DATA_CAP

Type Aliases

AddVerifierClientParams

AddVerifierParams

AllocationID

ClaimID

DataCap

DataCap is an integer number of bytes. We can introduce policy changes and replace this in the future.

ExtendClaimTermsReturn